refactor: files and folders renamed with respect to the naming convention for the project
