+++ lib/app/app.dart
@@ -1,6 +1,6 @@
 import 'package:rutorrentflutter/services/api/dev_api_service.dart';
 import 'package:rutorrentflutter/services/api/i_api_service.dart';
-import 'package:rutorrentflutter/theme/AppStateNotifier.dart';
+import 'package:rutorrentflutter/theme/app_state_notifier.dart';
 import 'package:rutorrentflutter/services/api/prod_api_service.dart';
 import 'package:rutorrentflutter/services/functional_services/authentication_service.dart';
 import 'package:rutorrentflutter/services/functional_services/disk_space_service.dart';
@@ -11,11 +11,11 @@ import 'package:rutorrentflutter/services/state_services/file_service.dart';
 import 'package:rutorrentflutter/services/state_services/history_service.dart';
 import 'package:rutorrentflutter/services/state_services/torrent_service.dart';
 import 'package:rutorrentflutter/services/state_services/user_preferences_service.dart';
-import 'package:rutorrentflutter/ui/views/DiskExplorer/disk_explorer_view.dart';
+import 'package:rutorrentflutter/ui/views/disk_explorer/disk_explorer_view.dart';
 import 'package:rutorrentflutter/ui/views/History/history_view.dart';
-import 'package:rutorrentflutter/ui/views/MediaPlayer/media_stream_view.dart';
+import 'package:rutorrentflutter/ui/views/media_player/media_stream_view.dart';
 import 'package:rutorrentflutter/ui/views/Settings/settings_view.dart';
-import 'package:rutorrentflutter/ui/views/TorrentDetail/torrent_detail_view.dart';
+import 'package:rutorrentflutter/ui/views/torrent_detail/torrent_detail_view.dart';
 import 'package:rutorrentflutter/utils/file_picker_service.dart';
 import 'package:stacked/stacked_annotations.dart';
 import 'package:stacked_services/stacked_services.dart';
