+++ lib/main.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:hive/hive.dart';
 import 'package:injectable/injectable.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:rutorrentflutter/theme/AppStateNotifier.dart';
-import 'package:rutorrentflutter/theme/AppTheme.dart';
+import 'package:rutorrentflutter/theme/app_state_notifier.dart';
+import 'package:rutorrentflutter/theme/app_theme.dart';
 import 'package:rutorrentflutter/app/app.locator.dart';
 import 'package:rutorrentflutter/app/app.router.dart';
 import 'package:rutorrentflutter/services/functional_services/notification_service.dart';
