+++ lib/ui/views/Settings/settings_view.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:rutorrentflutter/models/account.dart';
-import 'package:rutorrentflutter/theme/AppStateNotifier.dart';
+import 'package:rutorrentflutter/theme/app_state_notifier.dart';
 import 'package:rutorrentflutter/ui/views/Settings/settings_viewmodel.dart';
 import 'package:stacked/stacked.dart';
 
 class SettingsView extends StatelessWidget {
+  
   const SettingsView({Key? key}) : super(key: key);
 
   @override
