+++ lib/ui/views/rss_feed/rss_feed_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:rutorrentflutter/ui/views/RSSFeed/RSSFeed_viewmodel.dart';
-import 'package:rutorrentflutter/ui/widgets/smart_widgets/RSS%20Label%20Tile/rss_label_tile_view.dart';
+import 'package:rutorrentflutter/ui/views/rss_feed/rss_feed_viewmodel.dart';
+import 'package:rutorrentflutter/ui/widgets/smart_widgets/rss_label_tile/rss_label_tile_view.dart';
 import 'package:stacked/stacked.dart';
 
 class RSSFeedView extends StatelessWidget {
