+++ lib/ui/views/torrent_detail/torrent_detail_viewmodel.dart
