+++ lib/ui/widgets/dumb_widgets/show_disk_space_widget.dart
@@ -1,6 +1,6 @@
 import 'package:filesize/filesize.dart';
 import 'package:flutter/material.dart';
-import 'package:rutorrentflutter/theme/AppStateNotifier.dart';
+import 'package:rutorrentflutter/theme/app_state_notifier.dart';
 import 'package:rutorrentflutter/models/disk_space.dart';
 import 'package:rutorrentflutter/ui/shared/shared_styles.dart';
