+++ lib/ui/widgets/smart_widgets/rss_label_tile/rss_label_tile_view.dart
@@ -3,8 +3,8 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:intl/intl.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 import 'package:rutorrentflutter/models/rss.dart';
-import 'package:rutorrentflutter/ui/widgets/smart_widgets/RSS%20Label%20Tile/RSS%20Detail%20Sheet/rss_detail_sheet_view.dart';
-import 'package:rutorrentflutter/ui/widgets/smart_widgets/RSS%20Label%20Tile/rss_label_tile_viewmodel.dart';
+import 'package:rutorrentflutter/ui/widgets/smart_widgets/rss_label_tile/rss_detail_sheet/rss_detail_sheet_view.dart';
+import 'package:rutorrentflutter/ui/widgets/smart_widgets/rss_label_tile/rss_label_tile_viewmodel.dart';
 import 'package:stacked/stacked.dart';
 
 class RSSLabelTileView extends StatelessWidget {
