+++ lib/ui/widgets/smart_widgets/search_bar/search_bar_viewmodel.dart
