+++ lib/ui/widgets/smart_widgets/torrent_tile/torrent_tile_view.dart
@@ -1,10 +1,10 @@
 import 'package:filesize/filesize.dart';
 import 'package:flutter/material.dart';
-import 'package:rutorrentflutter/theme/AppStateNotifier.dart';
+import 'package:rutorrentflutter/theme/app_state_notifier.dart';
 import 'package:rutorrentflutter/models/torrent.dart';
 import 'package:rutorrentflutter/ui/shared/shared_styles.dart';
 import 'package:rutorrentflutter/ui/widgets/dumb_widgets/remove_torrent_dialog_widget.dart';
-import 'package:rutorrentflutter/ui/widgets/smart_widgets/Torrent%20Tile/torrent_tile_viewmodel.dart';
+import 'package:rutorrentflutter/ui/widgets/smart_widgets/torrent_tile/torrent_tile_viewmodel.dart';
 import 'package:stacked/stacked.dart';
 
 class TorrentTileView extends StatelessWidget {
